Move components into a src/components subdirection to match AFrame conventions
