+++ tests/src/components/outlineRenderer.test.js
@@ -1,9 +1,9 @@
 /* global assert, setup, suite, test */
 require('aframe');
-require('../../index');
-require('../../src/outlineRenderer');
+require('../../../index');
+require('../../../src/components/outlineRenderer');
 
-var entityFactory = require('../helpers').entityFactory;
+var entityFactory = require('../../helpers').entityFactory;
 
 var THREE = AFRAME.THREE;
